test(backup): add native/rclone backup benchmarking under read/write stress
Benchmark native/rClone backup  under read/write stress.
Test duration of backup to S3 and stress latencies.
Fixes: #10764
