+++ configurations/manager/2TB_backup_dataset.yaml
@@ -1,0 +1,9 @@
+
+# fill the cluster with data
+prepare_write_cmd: [ "cassandra-stress write cl=ALL n=536870912 -schema 'replication(strategy=NetworkTopologyStrategy,replication_factor=3)' -mode cql3 native  -rate threads=650 -col 'size=FIXED(1024) n=FIXED(1)' -pop seq=1..536870912",
+                     "cassandra-stress write cl=ALL n=536870912 -schema 'replication(strategy=NetworkTopologyStrategy,replication_factor=3)' -mode cql3 native  -rate threads=650 -col 'size=FIXED(1024) n=FIXED(1)' -pop seq=536870913..1073741824",
+                     "cassandra-stress write cl=ALL n=536870912 -schema 'replication(strategy=NetworkTopologyStrategy,replication_factor=3)' -mode cql3 native  -rate threads=650 -col 'size=FIXED(1024) n=FIXED(1)' -pop seq=1073741825..1610612736",
+                     "cassandra-stress write cl=ALL n=536870912 -schema 'replication(strategy=NetworkTopologyStrategy,replication_factor=3)' -mode cql3 native  -rate threads=650 -col 'size=FIXED(1024) n=FIXED(1)' -pop seq=1610612737..2147483648" ]
+
+# stress command to run
+stress_cmd_m: "cassandra-stress mixed cl=QUORUM duration=720m -schema 'replication(strategy=NetworkTopologyStrategy,replication_factor=3)' -mode cql3 native -rate 'threads=125 fixed=20000/s' -col 'size=FIXED(1024) n=FIXED(1)' -pop 'dist=gauss(1..2000000000,1000000000,300000000)'"
