+++ configurations/manager/manager_native_backup_nemesis.yaml
@@ -1,0 +1,1 @@
+nemesis_class_name: 'ManagerNativeBackup'
