+++ configurations/manager/manager_rclone_backup_nemesis.yaml
@@ -1,0 +1,1 @@
+nemesis_class_name: 'ManagerRcloneBackup'
