+++ data_dir/nemesis.yml
@@ -478,6 +478,22 @@ disrupt_major_compaction:
   supports_high_disk_utilization: true
   topology_changes: false
   zero_node_changes: false
+disrupt_manager_backup:
+  config_changes: false
+  delete_rows: false
+  disruptive: false
+  free_tier_set: false
+  has_steady_run: false
+  kubernetes: false
+  limited: false
+  manager_operation: true
+  networking: false
+  run_with_gemini: true
+  schema_changes: false
+  sla: false
+  supports_high_disk_utilization: false
+  topology_changes: false
+  zero_node_changes: false
 disrupt_maximum_allowed_sls_with_max_shares_during_load:
   config_changes: false
   delete_rows: false
