+++ data_dir/nemesis_classes.yml
@@ -34,6 +34,8 @@
 - IsolateNodeWithProcessSignalNemesis
 - LoadAndStreamMonkey
 - MajorCompactionMonkey
+- ManagerNativeBackup
+- ManagerRcloneBackup
 - MemoryStressMonkey
 - MgmtBackup
 - MgmtBackupSpecificKeyspaces
