+++ performance_regression_test.py
@@ -16,6 +16,7 @@
 
 import os
 import time
+from typing import Optional
 
 import yaml
 from cassandra.query import SimpleStatement
@@ -500,6 +501,15 @@ class PerformanceRegressionTest(ClusterTester, loader_utils.LoaderUtilsMixin):
                     AND speculative_retry = 'NONE';
             """)
 
+    @optional_stage('perf_steady_state_calc')
+    @latency_calculator_decorator
+    def steady_state_latency(self, hdr_tags: list[str], sleep_time: Optional[int] = None):
+        # NOTE: 'hdr_tags' will be used by the 'latency_calculator_decorator' decorator
+        sleep_time = sleep_time or self.db_cluster.params.get('nemesis_interval') * 60
+        InfoEvent(message='Starting Steady State calculation for %ss' % sleep_time).publish()
+        time.sleep(sleep_time)
+        InfoEvent(message='Ended Steady State calculation. Took %ss' % sleep_time).publish()
+
     # Base Tests
     def test_write(self):
         """
@@ -861,15 +871,6 @@ class PerformanceRegressionUpgradeTest(PerformanceRegressionTest, UpgradeTest):
                                          event_class=CassandraStressEvent,
                                          extra_time_to_expiration=60):
             self.loaders.kill_stress_thread()
-
-    @optional_stage('perf_steady_state_calc')
-    @latency_calculator_decorator
-    def steady_state_latency(self, hdr_tags: list[str]):
-        # NOTE: 'hdr_tags' will be used by the 'latency_calculator_decorator' decorator
-        sleep_time = self.db_cluster.params.get('nemesis_interval') * 60
-        InfoEvent(message='Starting Steady State calculation for %ss' % sleep_time).publish()
-        time.sleep(sleep_time)
-        InfoEvent(message='Ended Steady State calculation. Took %ss' % sleep_time).publish()
 
     @latency_calculator_decorator
     def post_upgrades_steady_state(self, hdr_tags: list[str]):
