+++ unit_tests/test_scylla_yaml.py
@@ -412,7 +412,8 @@ class ScyllaYamlTest(unittest.TestCase):
                 'tablets_mode_for_new_keyspaces': None,
                 'force_gossip_topology_changes': None,
                 'reader_concurrency_semaphore_cpu_concurrency': None,
-                'vector_store_uri': None
+                'vector_store_uri': None,
+                'object_storage_endpoints': None,
             }
         )
